Fix join table bigint->integer, migrate db

+++ db/migrate/20200827191522_create_countries_target_groups.rb
@@ -1,8 +1,8 @@
 class CreateCountriesTargetGroups < ActiveRecord::Migration[6.0]
   def change
     create_table :countries_target_groups, id: false do |t|
-      t.bigint :country_id
-      t.bigint :target_group_id
+      t.integer :country_id
+      t.integer :target_group_id
     end
     
     add_index :countries_target_groups, :country_id
